Only delete old docs directory if it exists

+++ dev/tools/docs/build_html_brian2.py
@@ -2,5 +2,6 @@ import os
 import shutil
 
 os.chdir('../../../docs_sphinx')
-shutil.rmtree('../docs')
+if os.path.exists('../docs'):
+    shutil.rmtree('../docs')
 os.system('sphinx-build -b html . ../docs')
